# Conflicts:
#	WordPress/build.gradle
#	WordPress/src/main/java/org/wordpress/android/ui/media/MediaPreviewFragment.java

+++ WordPress/src/main/java/org/wordpress/android/ui/stories/StoryComposerViewModel.kt
@@ -10,6 +10,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModel
+import com.wordpress.stories.compose.frame.StorySaveEvents.StorySaveResult
+import org.wordpress.android.viewmodel.SingleLiveEvent
 import org.wordpress.android.WordPress
 import org.wordpress.android.fluxc.Dispatcher
 import org.wordpress.android.fluxc.generated.PostActionBuilder
@@ -21,9 +23,14 @@ import org.wordpress.android.ui.notifications.SystemNotificationsTracker
 import org.wordpress.android.ui.posts.EditPostRepository
 import org.wordpress.android.ui.posts.PostEditorAnalyticsSession
 import org.wordpress.android.ui.posts.PostEditorAnalyticsSession.Outcome.CANCEL
+import org.wordpress.android.ui.posts.PostEditorAnalyticsSession.Outcome.PUBLISH
+import org.wordpress.android.ui.posts.PostEditorAnalyticsSession.Outcome.SAVE
 import org.wordpress.android.ui.posts.PostEditorAnalyticsSessionWrapper
 import org.wordpress.android.ui.posts.SavePostToDbUseCase
+import org.wordpress.android.ui.stories.StoryComposerActivity.Companion.STATE_KEY_ORIGINAL_STORY_SAVE_RESULT
 import org.wordpress.android.ui.stories.usecase.SetUntitledStoryTitleIfTitleEmptyUseCase
+import org.wordpress.android.util.AppLog
+import org.wordpress.android.util.AppLog.T
 import org.wordpress.android.util.helpers.MediaFile
 import org.wordpress.android.viewmodel.Event
 import javax.inject.Inject
@@ -43,16 +50,17 @@ class StoryComposerViewModel @Inject constructor(
 
     private lateinit var editPostRepository: EditPostRepository
     private lateinit var site: SiteModel
-    private lateinit var postEditorAnalyticsSession: PostEditorAnalyticsSession
+    private var postEditorAnalyticsSession: PostEditorAnalyticsSession? = null
+    private var originalIntentStorySaveResult: StorySaveResult? = null
 
     private val _mediaFilesUris = MutableLiveData<List<Uri>>()
     val mediaFilesUris: LiveData<List<Uri>> = _mediaFilesUris
 
-    private val _openPrepublishingBottomSheet = MutableLiveData<Event<Unit>>()
-    val openPrepublishingBottomSheet: LiveData<Event<Unit>> = _openPrepublishingBottomSheet
+    private val _openPrepublishingBottomSheet = SingleLiveEvent<Event<Unit>>()
+    val openPrepublishingBottomSheet = _openPrepublishingBottomSheet
 
-    private val _submitButtonClicked = MutableLiveData<Event<Unit>>()
-    val submitButtonClicked: LiveData<Event<Unit>> = _submitButtonClicked
+    private val _submitButtonClicked = SingleLiveEvent<Event<Unit>>()
+    val submitButtonClicked = _submitButtonClicked
 
     init {
         lifecycleOwner.lifecycleRegistry.currentState = Lifecycle.State.CREATED
@@ -61,15 +69,22 @@ class StoryComposerViewModel @Inject constructor(
     private val _trackEditorCreatedPost = MutableLiveData<Event<Unit>>()
     val trackEditorCreatedPost: LiveData<Event<Unit>> = _trackEditorCreatedPost
 
+    @Suppress("LongParameterList")
     fun start(
         site: SiteModel,
         editPostRepository: EditPostRepository,
         postId: LocalId,
         postEditorAnalyticsSession: PostEditorAnalyticsSession?,
-        notificationType: NotificationType?
-    ) {
+        notificationType: NotificationType?,
+        originalStorySaveResult: StorySaveResult?
+    ): Boolean {
         this.editPostRepository = editPostRepository
         this.site = site
+        this.originalIntentStorySaveResult = originalStorySaveResult
+
+        notificationType?.let {
+            systemNotificationsTracker.trackTappedNotification(it)
+        }
 
         if (postId.value == 0) {
             // Create a new post
@@ -80,14 +95,18 @@ class StoryComposerViewModel @Inject constructor(
             editPostRepository.loadPostByLocalPostId(postId.value)
         }
 
-        setupPostEditorAnalyticsSession(postEditorAnalyticsSession)
-
-        notificationType?.let {
-            systemNotificationsTracker.trackTappedNotification(it)
+        // Ensure we have a valid post
+        if (!editPostRepository.hasPost()) {
+            AppLog.e(T.EDITOR, "StoryComposerViewModel's EditPostRepository has no Post loaded: " + postId.value)
+            return false
         }
+
+        setupPostEditorAnalyticsSession(postEditorAnalyticsSession)
 
         lifecycleOwner.lifecycleRegistry.currentState = Lifecycle.State.STARTED
         updateStoryPostWithChanges()
+
+        return true
     }
 
     private fun setupPostEditorAnalyticsSession(postEditorAnalyticsSession: PostEditorAnalyticsSession?) {
@@ -95,7 +114,7 @@ class StoryComposerViewModel @Inject constructor(
                 editPostRepository.getPost(),
                 site
         )
-        this.postEditorAnalyticsSession.start(null)
+        this.postEditorAnalyticsSession?.start(null, null, null)
     }
 
     private fun createPostEditorAnalyticsSessionTracker(
@@ -112,12 +131,25 @@ class StoryComposerViewModel @Inject constructor(
         outState.putSerializable(WordPress.SITE, site)
         outState.putInt(StoryComposerActivity.STATE_KEY_POST_LOCAL_ID, editPostRepository.id)
         outState.putSerializable(StoryComposerActivity.STATE_KEY_EDITOR_SESSION_DATA, postEditorAnalyticsSession)
+        outState.putParcelable(STATE_KEY_ORIGINAL_STORY_SAVE_RESULT, originalIntentStorySaveResult)
     }
 
-    fun onStoryDiscarded() {
-        // delete empty post from database
-        dispatcher.dispatch(PostActionBuilder.newRemovePostAction(editPostRepository.getEditablePost()))
-        postEditorAnalyticsSession.setOutcome(CANCEL)
+    fun onStorySaved() {
+        postEditorAnalyticsSession?.setOutcome(SAVE)
+    }
+
+    // returns true if user is discarding a Story out of a save retry (error handling state)
+    // returns false otherwise
+    fun onStoryDiscarded(deleteDiscardedPost: Boolean): Boolean {
+        if (deleteDiscardedPost) {
+            // delete empty post from database
+            dispatcher.dispatch(PostActionBuilder.newRemovePostAction(editPostRepository.getEditablePost()))
+        }
+        postEditorAnalyticsSession?.setOutcome(CANCEL)
+
+        originalIntentStorySaveResult?.let {
+            return (!it.isSuccess() || it.isRetry)
+        } ?: return false
     }
 
     private fun updateStoryPostWithChanges() {
@@ -146,12 +178,13 @@ class StoryComposerViewModel @Inject constructor(
 
     fun onSubmitButtonClicked() {
         setUntitledStoryTitleIfTitleEmptyUseCase.setUntitledStoryTitleIfTitleEmpty(editPostRepository)
+        postEditorAnalyticsSession?.setOutcome(PUBLISH)
         _submitButtonClicked.postValue(Event(Unit))
     }
 
     override fun onCleared() {
         super.onCleared()
         lifecycleOwner.lifecycleRegistry.currentState = Lifecycle.State.DESTROYED
-        postEditorAnalyticsSession.end()
+        postEditorAnalyticsSession?.end(null)
     }
 }
